Put reboot wrapper in separate file so it can be used from other test files. Apply wrapper to EEPROM stress fn.
